Uncommited changes

+++ install.sh
@@ -1,11 +1,12 @@
 #!/bin/sh
 
 
-rm -rf /opt/octotouch/
 mkdir -p /opt/octotouch
 mkdir -p /opt/octotouch/css
 mkdir -p /opt/octotouch/glade
-cp build/octotouch.cfg /opt/octotouch/
+if [ ! -f /opt/octotouch/octotouch.cfg ]; then
+	cp build/octotouch.cfg /opt/octotouch/
+fi
 cp build/octotouch /opt/octotouch/
 cp build/css/* /opt/octotouch/css
 cp build/glade/* /opt/octotouch/glade
